#67: Inited a new embedding router
